Refactor: split into smaller files + exception management + minors fix
