+++ bindings/Octave/CMakeLists.txt
@@ -11,9 +11,10 @@ set(LIBKRIGING_SOURCE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/../..")
 include(${LIBKRIGING_SOURCE_DIR}/cmake/configureOctave.cmake)
 
 add_subdirectory(tools)
-add_subdirectory(toys)
 
-add_mex_function(NAME mLibKriging SOURCES mLibKriging.cpp LINK_LIBRARIES Kriging OctaveShared OctaveToys)
+add_mex_function(NAME mLibKriging
+        SOURCES mLibKriging.cpp LinearRegression_binding.cpp LinearRegression_binding.hpp 
+        LINK_LIBRARIES Kriging OctaveShared)
 add_mex_function(NAME mypow2 SOURCES mypow2.cpp)
 add_mex_function(NAME mkStructure SOURCES mkStructure.cpp)
 add_mex_function(NAME demoStruct SOURCES demoStruct.cpp)
