+++ bindings/Octave/LinearRegression.m
@@ -1,6 +1,6 @@
 classdef LinearRegression < handle
     % attributes doc: https://fr.mathworks.com/help/matlab/matlab_oop/property-attributes.html
-    properties (NonCopyable, Transient) % Access = protected, Hidden, 
+    properties (NonCopyable, Transient, Access = protected, Hidden) 
         ref
     end
 
@@ -10,20 +10,20 @@ classdef LinearRegression < handle
             mLibKriging("LinearRegression::new", obj, varargin{:});
         end
         
-        function delete(obj)
+        function delete(obj, varargin)
             % destroy the mex backend
             if ~isempty(obj.ref)
                 printf("Delete LinearRegression\n")
-                mLibKriging("LinearRegression::delete",obj)
+                mLibKriging("LinearRegression::delete",obj, varargin{:})
             end
         end
         
-        function fit(obj, y, X)
-            mLibKriging("LinearRegression::fit", obj, y, X);
+        function fit(obj, varargin)
+            mLibKriging("LinearRegression::fit", obj, varargin{:});
         end
         
-        function varargout = predict(obj, X)
-            [varargout{1:nargout}] = mLibKriging("LinearRegression::predict", obj, X);
+        function varargout = predict(obj, varargin)
+            [varargout{1:nargout}] = mLibKriging("LinearRegression::predict", obj, varargin{:});
         end
         
     end
