+++ bindings/Octave/tools/CMakeLists.txt
@@ -1,8 +1,23 @@
-add_library(OctaveShared SHARED
-        ObjectCollector.cpp ObjectCollector.hpp)
+add_library(OctaveShared
+        NonCopyable.hpp
+        string_hash.hpp
+        ObjectCollector.cpp ObjectCollector.hpp
+        relative_error.cpp relative_error.hpp
+        ObjectAccessor.cpp ObjectAccessor.hpp
+        RequiresArg.cpp RequiresArg.hpp
+        mx_accessor.hpp
+        overload.hpp
+        MxMapper.cpp MxMapper.hpp
+        MxException.hpp 
+        formatString.hpp)
 
-if (CXX_CLANG_TIDY)
-    set_target_properties(OctaveShared
-            PROPERTIES
-            CXX_CLANG_TIDY ${CXX_CLANG_TIDY})
-endif ()
+target_link_libraries(OctaveShared armadillo)
+
+set_target_properties(OctaveShared PROPERTIES
+        COMPILE_FLAGS "${OCT_CPPFLAGS} ${OCT_CXXPICFLAGS} ${OCT_CXXFLAGS}")
+
+#if (CXX_CLANG_TIDY)
+#    set_target_properties(OctaveShared
+#            PROPERTIES
+#            CXX_CLANG_TIDY ${CXX_CLANG_TIDY})
+#endif ()
