+++ bindings/Octave/tools/NonCopyable.hpp
@@ -1,7 +1,3 @@
-//
-// Created by Pascal Havé on 07/04/2020.
-//
-
 #ifndef LIBKRIGING_BINDINGS_OCTAVE_TOOLS_NONCOPYABLE_HPP
 #define LIBKRIGING_BINDINGS_OCTAVE_TOOLS_NONCOPYABLE_HPP
