+++ bindings/Octave/tools/relative_error.hpp
@@ -1,7 +1,3 @@
-//
-// Created by Pascal Havé on 08/04/2020.
-//
-
 #ifndef LIBKRIGING_BINDINGS_OCTAVE_TOOLS_RELATIVE_ERROR_HPP
 #define LIBKRIGING_BINDINGS_OCTAVE_TOOLS_RELATIVE_ERROR_HPP
