--- bindings/Octave/toys/func1.hpp
@@ -1,10 +1,0 @@
-//
-// Created by Pascal Havé on 08/04/2020.
-//
-
-#ifndef LIBKRIGING_BINDINGS_OCTAVE_TOYS_FUNC1_HPP
-#define LIBKRIGING_BINDINGS_OCTAVE_TOYS_FUNC1_HPP
-
-#include <mex.h>
-void func1(int nlhs, mxArray* plhs[], int nrhs, const mxArray* prhs[]);
-#endif  // LIBKRIGING_BINDINGS_OCTAVE_TOYS_FUNC1_HPP
